+++ my-react-app/package.json
@@ -13,6 +13,7 @@
     "firebase": "^10.13.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-icons": "^5.3.0",
     "react-router-dom": "^6.26.2"
   },
   "devDependencies": {
